Separate out remote relation workers to separate files
